requirements.txt
aiofiles==0.4.0
aiohttp==3.4.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
beautifulsoup4==4.6.3
cached-property==1.5.1
chardet==3.0.4
configparser==3.5.0
idna==2.7
idna-ssl==1.1.0
lxml==4.2.5
more-itertools==4.3.0
multidict==4.4.2
six==1.11.0
studip-api==2.2
yarl==1.2.6
